+++ contrib.json
@@ -202,8 +202,11 @@
         [ "git pull upstream stable",                "Update the release branch" ],
         [ "npm install",                             "Ensure dependency updates have been installed" ],
         [ "grunt test",                              "Run tests" ],
-        [ "grunt chg-release:{{ release_type }}",    "Update the changelog with the new release" ],
         [ "grunt version:{{ release_type }}",        "Bump package versions" ],
+        { "exec": "grunt vjs-version",
+          "id": "version",                   "desc": "Return the current VJS Version from the package.json file"},
+        [ "grunt chg-release:{{ version }}",         "Update the changelog with the new release" ],
+        [ "grunt clean:dist",                        "Clean out the dist folder before the build" ],
         [ "grunt",                                   "Build the release" ],
         [ "git add dist/video-js --force",           "Add the (otherwise ignored) release files" ],
         [ "grunt vjsdocs",                           "Rebuild the docs" ],
@@ -212,10 +215,11 @@
         [ "git push upstream stable",                "Push changes to the remote" ],
         [ "git push upstream --tags",                "Push tags to the remote" ],
         [ "npm publish",                             "Publish to npm" ],
+        [ "grunt github-release",                    "Create a new release on Github"],
         [ "git checkout master",                     "Checkout the development branch" ],
         [ "git pull upstream master",                "Update the development branch" ],
         [ "git merge stable",                        "Merge changes" ],
-        [ "grunt clean:dist",                        "Remove built dist files to re-ignore them" ],
+        [ "grunt clean:dist",                        "Remove built dist files again to re-ignore them" ],
         [ "git add -u",                              "Add the changes for the removed dist files" ],
         [ "git commit -m 'Removed dist files'",      "Commit the removed dist files" ],
         [ "git push upstream master",                "Push development branch changes" ]
