+++ package.json
@@ -26,7 +26,7 @@
   },
   "devDependencies": {
     "calcdeps": "~0.1.7",
-    "chg": "~0.1.8",
+    "chg": "~0.2.0",
     "contribflow": "~0.2.0",
     "github": "~0.1.14",
     "grunt": "^0.4.4",
@@ -40,10 +40,13 @@
     "grunt-contrib-less": "~0.6.4",
     "grunt-contrib-qunit": "~0.2.1",
     "grunt-contrib-watch": "~0.1.4",
+    "grunt-fastly": "^0.1.3",
+    "grunt-github-releaser": "^0.1.17",
     "grunt-karma": "^0.8.3",
     "grunt-s3": "~0.2.0-alpha",
     "grunt-tagrelease": "~0.3.3",
     "grunt-version": "~0.3.0",
+    "grunt-videojs-languages": "0.0.4",
     "grunt-zip": "0.10.2",
     "karma": "^0.12.14",
     "karma-chrome-launcher": "^0.1.3",
@@ -58,10 +61,8 @@
     "mocha": "~1.8.1",
     "open": "0.0.4",
     "qunitjs": "~1.14.0",
-    "semver": "~1.1.4",
     "sinon": "~1.9.1",
     "uglify-js": "~2.3.6",
-    "videojs-doc-generator": "0.0.1",
-    "grunt-videojs-languages": "0.0.4"
+    "videojs-doc-generator": "0.0.1"
   }
 }
