+++ tasks/vjs-version.js
@@ -1,0 +1,7 @@
+module.exports = function(grunt) {
+  grunt.registerTask('vjs-version', function() {
+    var version = grunt.vjsVersion.full;
+    grunt.log.writeln(version);
+    return version;
+  });
+}
